[feat] add distance to hotelstar

+++ gaenolja/src/main/java/com/gaenolja/model/dto/Hotel.java
@@ -57,13 +57,13 @@ public class Hotel {
 	public double getLatitude() {
 		return latitude;
 	}
-	public void setLatitude(long latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
 	public double getLongitude() {
 		return longitude;
 	}
-	public void setLongitude(long longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
 	public String getAddress() {
